Change source directory and script name
