--- vue/components/NotFound.vue
@@ -1,20 +1,0 @@
-<template>
-    <h2>{{ t('The page was not found') }}</h2>
-</template>
-
-<script>
-export default {
-    name: 'NotFound',
-    mounted () {
-        this.$store.dispatch('setPage', { page: 'notfound' })
-    }
-}
-</script>
-
-<style scoped>
-
-h2 {
-    margin: 1rem;
-}
-
-</style>
